+++ R/partition.R
@@ -40,7 +40,7 @@
 bqr_partition <- function(sharded,
                           partition,
                           projectId = bq_get_global_project(),
-                          datasetId){
+                          datasetId = bq_get_global_dataset()){
   
   ## check for shared tables
   tables <- bqr_list_tables(projectId = projectId, datasetId = datasetId)