+++ R/query.R
@@ -31,7 +31,7 @@
 #' @family BigQuery query functions
 #' @export
 bqr_query <- function(projectId = bq_get_global_project(), 
-                      datasetId, 
+                      datasetId = bq_get_global_dataset(), 
                       query, 
                       maxResults = 1000, 
                       useLegacySql = TRUE, 
@@ -178,7 +178,7 @@ bqr_query <- function(projectId = bq_get_global_project(),
 #' @family BigQuery asynch query functions  
 #' @export
 bqr_query_asynch <- function(projectId = bq_get_global_project(), 
-                             datasetId, 
+                             datasetId = bq_get_global_dataset(), 
                              query, 
                              destinationTableId,
                              useLegacySql = TRUE,