+++ R/tables.R
@@ -16,7 +16,8 @@
 #' @family bigQuery meta functions
 #' @export
 bqr_list_tables <- function(projectId = bq_get_global_project(), 
-                            datasetId, maxResults = 1000, pageToken = ""){
+                            datasetId = bq_get_global_dataset(), 
+                            maxResults = 1000, pageToken = ""){
   
   l <- googleAuthR::gar_api_generator("https://www.googleapis.com/bigquery/v2",
                                       "GET",
@@ -64,7 +65,7 @@ parse_bqr_list_tables <- function(x) {
 #' 
 #' @family bigQuery meta functions
 #' @export
-bqr_table_meta <- function(projectId = bq_get_global_project(), datasetId, tableId){
+bqr_table_meta <- function(projectId = bq_get_global_project(), datasetId = bq_get_global_dataset(), tableId){
   
   
   f <- function(x){
@@ -98,7 +99,7 @@ bqr_table_meta <- function(projectId = bq_get_global_project(), datasetId, table
 #' 
 #' @family bigQuery meta functions
 #' @export
-bqr_table_data <- function(projectId = bq_get_global_project(), datasetId, tableId,
+bqr_table_data <- function(projectId = bq_get_global_project(), datasetId = bq_get_global_dataset(), tableId,
                            maxResults = 1000){
   
   l <- googleAuthR::gar_api_generator("https://www.googleapis.com/bigquery/v2",
@@ -139,7 +140,7 @@ bqr_table_data <- function(projectId = bq_get_global_project(), datasetId, table
 #' @family bigQuery meta functions
 #' @export
 bqr_create_table <- function(projectId = bq_get_global_project(), 
-                             datasetId, 
+                             datasetId = bq_get_global_dataset(), 
                              tableId, 
                              template_data,
                              timePartitioning = FALSE,
@@ -207,7 +208,7 @@ bqr_create_table <- function(projectId = bq_get_global_project(),
 #' 
 #' @family bigQuery meta functions
 #' @export
-bqr_delete_table <- function(projectId = bq_get_global_project(), datasetId, tableId){
+bqr_delete_table <- function(projectId = bq_get_global_project(), datasetId = bq_get_global_dataset(), tableId){
   
   l <- googleAuthR::gar_api_generator("https://www.googleapis.com/bigquery/v2",
                                       "DELETE",